Moves the logic code from pokefly to fragments
