+++ app/src/main/java/com/kamron/pogoiv/Fragments/MovesFragment.java
@@ -6,8 +6,16 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 
+import com.kamron.pogoiv.Pokefly;
 import com.kamron.pogoiv.R;
+import com.kamron.pogoiv.scanlogic.PokemonShareHandler;
+import com.kamron.pogoiv.scanlogic.ScanContainer;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+import butterknife.OnClick;
 
 
 /**
@@ -15,17 +23,32 @@ import com.kamron.pogoiv.R;
  */
 public class MovesFragment extends Fragment {
 
+    @BindView(R.id.shareWithStorimod)
+    ImageView shareWithStorimod;
+
+    Pokefly pokefly;
 
     public MovesFragment() {
-        // Required empty public constructor
+
     }
+
 
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_moves, container, false);
+        View thisView = inflater.inflate(R.layout.fragment_moves, container, false);
+        ButterKnife.bind(this, thisView);
+        return thisView;
     }
 
+    @OnClick({R.id.shareWithOtherApp})
+    /**
+     * Creates an intent to share the result of the pokemon scan, and closes the overlay.
+     */
+    public void shareScannedPokemonInformation() {
+        PokemonShareHandler communicator = new PokemonShareHandler();
+        communicator.spreadResultIntent(pokefly, ScanContainer.scanContainer.currScan, pokefly.pokemonUniqueID);
+        pokefly.cancelInfoDialog();
+    }
 }
